pyproject.toml
[tool.poetry]
name = "thesis"
version = "0.1.0"
description = "AI for AI - automatic podcast generator for AI news"
authors = ["Beatriz Barretto <biachedid@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
numpy = "<=2.1"
feedparser = "^6.0.11"
pandas = "^2.2.3"
bs4 = "^0.0.2"
pydantic = "^2.10.6"
markdownify = "^0.14.1"
requests = "^2.32.3"
sentence-transformers = "^3.4.1"
bertopic = "^0.16.4"
llvmlite = "^0.44.0"
nltk = "^3.9.1"
numba = "^0.61.0"
scipy = ">=1.10,<1.14"
datasets = "^3.3.2"
jupyter = "^1.1.1"
transformers = {extras = ["torch"], version = "^4.49.0"}
hdbscan = "^0.8.33"
scikit-learn = "^1.3"
gensim = "^4.3.3"
bert-score = "^0.3.13"
googlesearch-python = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
